refactor bus service, upload to remote server

+++ src/main/java/org/ahmedukamel/eduai/service/bus/IBusManagementService.java
@@ -9,6 +9,5 @@ public interface IBusManagementService {
 
     Object getBus(Long id);
 
-    Object getAllBuses(Integer schoolId, long pageSize, long pageNumber);
-
+    Object getAllBuses(Integer schoolId, int pageSize, int pageNumber);
 }
